refactor(*): ディレクトリ構成の修正
* `src/`がごちゃごちゃしていたので，`__tests__`，`components`，`pages`ごとにファイルを整理
* それに応じて`import`文を修正
