+++ src/__tests__/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen, test, expect } from "@testing-library/react";
-import App from "./App";
+import App from "../pages/App";
 
 test("renders learn react link", () => {
   render(<App />);
