+++ src/pages/SignUp.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { initializeApp } from "firebase/app";
 import { getAuth, createUserWithEmailAndPassword } from "firebase/auth";
-import { DevDaemonDBController } from "./DevDaemonDBController.ts";
-import { firebaseConfig } from "./FirebaseConfig";
-import { useAuthContext } from "./AuthContext";
+import { DevDaemonDBController } from "../DevDaemonDBController.ts";
+import { firebaseConfig } from "../FirebaseConfig";
+import { useAuthContext } from "../AuthContext";
 import { useHistory } from "react-router-dom";
-import FormBox from "./FormBox";
-import Button from "./Button";
+import FormBox from "../components/FormBox";
+import Button from "../components/Button";
 import "./SignUp.css";
 
 function SignUp() {
